requirements.txt
numpy
matplotlib
torch
jpype1
POT